# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Real Estate Search API",
    "description": "\nNatural language search for Romanian real estate with conversation memory.\n\n## Features\n- 🔍 Natural language queries in Romanian or English\n- 💬 Conversation memory - follow-up queries refine search\n- 🔐 Keycloak JWT authentication\n- 🛡️ Document Level Security - users only see their own sessions\n- 🏷️ Rich result data for UI cards (images, attributes, etc.)\n\n## Authentication\nPass your Keycloak JWT via `Authorization: Bearer <token>` header.\nSet `AUTH_ENABLED=true` to require authentication.\n\n## Examples\n- \"Caut apartament de inchiriat in Pallady cu 2 camere\"\n- \"sa fie pet friendly\" (follow-up)\n- \"pret maxim 600 euro\" (follow-up)\n    ",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "description": "Health check endpoint",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/search": {
      "post": {
        "summary": "Search Listings",
        "description": "Search real estate listings using natural language.\n\nThe query is parsed by an LLM to extract structured filters.\nConversation context is maintained per session for follow-up queries.",
        "operationId": "search_listings_search_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Session ID for conversation continuity",
              "title": "Session Id"
            },
            "description": "Session ID for conversation continuity"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/{session_id}": {
      "get": {
        "summary": "Get Session",
        "description": "Get current session state (DLS: own sessions only)",
        "operationId": "get_session_session__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Session Endpoint",
        "description": "Delete a session (DLS: own sessions only)",
        "operationId": "delete_session_endpoint_session__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/{session_id}/history": {
      "get": {
        "summary": "Get Session History",
        "description": "Get query history for a session (DLS: own sessions only)",
        "operationId": "get_session_history_session__session_id__history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryItem"
                  },
                  "title": "Response Get Session History Session  Session Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/{session_id}/reset": {
      "post": {
        "summary": "Reset Session",
        "description": "Reset session filters but keep session ID",
        "operationId": "reset_session_session__session_id__reset_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions": {
      "get": {
        "summary": "List User Sessions",
        "description": "List all sessions for current user (DLS enforced)",
        "operationId": "list_user_sessions_sessions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "summary": "Get Current User Info",
        "description": "Get information about current authenticated user",
        "operationId": "get_current_user_info_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryItem": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "query",
          "timestamp"
        ],
        "title": "HistoryItem",
        "description": "Single query history item"
      },
      "SearchFeatures": {
        "properties": {
          "animale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Animale"
          },
          "fumatori": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fumatori"
          },
          "parcare": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parcare"
          },
          "mobilat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mobilat"
          },
          "centrala": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Centrala"
          }
        },
        "type": "object",
        "title": "SearchFeatures",
        "description": "Feature preferences (pet friendly, parking, etc.)"
      },
      "SearchFilters": {
        "properties": {
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "transaction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction"
          },
          "property_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Type"
          },
          "rooms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rooms"
          },
          "price_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Min"
          },
          "price_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Max"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords",
            "default": []
          },
          "features": {
            "$ref": "#/components/schemas/SearchFeatures"
          },
          "exclude_agencies": {
            "type": "boolean",
            "title": "Exclude Agencies",
            "description": "Filter out known agents/agencies",
            "default": false
          }
        },
        "type": "object",
        "title": "SearchFilters",
        "description": "Parsed search filters from LLM"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Natural language search query"
          },
          "size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Size",
            "description": "Number of results",
            "default": 25
          },
          "exclude_agencies": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclude Agencies",
            "description": "UI toggle: true=hide agencies, false=show all, null=use query parsing"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest",
        "description": "Search request with natural language query"
      },
      "SearchResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "parsed_filters": {
            "$ref": "#/components/schemas/SearchFilters"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "query",
          "parsed_filters",
          "total",
          "results",
          "session_id",
          "user_id"
        ],
        "title": "SearchResponse",
        "description": "Full search response"
      },
      "SearchResult": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "OpenSearch document ID"
          },
          "ad_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Id",
            "description": "Original ad ID from source"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Truncated description",
            "default": ""
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          },
          "location": {
            "type": "string",
            "title": "Location",
            "description": "Location: City, Area",
            "default": ""
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories",
            "description": "Tags: Inchiriere, Apartamente, 2 camere"
          },
          "surface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "image_count": {
            "type": "integer",
            "title": "Image Count",
            "default": 0
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "score": {
            "type": "integer",
            "title": "Score",
            "description": "Relevance score 0-100 for UI indicator",
            "default": 0
          },
          "is_agency": {
            "type": "boolean",
            "title": "Is Agency",
            "description": "True if seller is known agent/agency",
            "default": false
          },
          "seller_type": {
            "type": "string",
            "title": "Seller Type",
            "description": "private, agent, agency, or unknown",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "SearchResult",
        "description": "Single search result - streamlined for card UI"
      },
      "SessionInfo": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "filters": {
            "$ref": "#/components/schemas/SearchFilters"
          },
          "query_count": {
            "type": "integer",
            "title": "Query Count"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "session_id",
          "filters",
          "query_count"
        ],
        "title": "SessionInfo",
        "description": "Session state information"
      },
      "UserInfo": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "default": []
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Groups",
            "default": []
          },
          "is_anonymous": {
            "type": "boolean",
            "title": "Is Anonymous",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "UserInfo",
        "description": "Current user information"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}